training/requirements.txt
tensorflow>=2.15,<2.18
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
scikit-learn>=1.3
